Include symbols in the Currencies database and add query access to it from the BudgetProvider
This commit modifies the BudgetContract.CurrenciesTable class so that we now store a
custom value for the symbol, rather than relying on javas Currency implementation, which
sometimes provides inconsistent symbols.

When the table is created for the first time, it gets populated by some hard-coded
currency values. This is because I don't want to bother downloading the supported
currencies; they probably won't change very much.

It also then provides a ContentProvider URI and query access in the BudgetProvider, so that
the business logic can access it all.