Create a different database for each user that signs into the app
This should reduce the complexity of having to store a user's table and having
a constraint with every other piece of information in the database.

I still need to extract some of the common data, like exchange rates and currency
information, which will be common to each user. In this 'common' database, I could
also have a user's table, storing their key information, which would allow offline
log in for several users.

Furthermore, I also created an "ENTER_APP" page in the launch fragment, which is
to be used instead of passing null to enter the main app. This should help improve
readability.